+++ database/migrations/2020_08_02_113710_remove_wallet_owner_from_money_transactions.php
@@ -1,0 +1,32 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class RemoveWalletOwnerFromMoneyTransactions extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('money_transactions', function (Blueprint $table) {
+            $table->dropColumn('wallet_owner');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('money_transactions', function (Blueprint $table) {
+            $table->string('wallet_owner')->after('description')->nullable();
+        });
+    }
+}