+++ resources/lang/en/accounting.php
@@ -27,7 +27,6 @@ return [
     'no_receipt' => 'No receipt',
     'difference' => 'Difference',
     'month' => 'Month',
-    'wallet_owner' => 'Wallet owner',
     'confirm_delete_receipt' => 'Really delete this receipt?',
     'selection_for_webling' => 'Selection for Webling Accounting',
     'credit' => 'Credit',