+++ Changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## NEXT
+
+* Remove wallet owner field from money transactions in accounting module
+
 ## 2.23.0
 
 * Added global wallet summary